Add support for drawing in the plot
Left-click-drag to draw the first poly-line segment into the plot. New
secions is added by clicking or dragging. To start a new line press one of
Enter, Space or ESC. To delete the last segment press DEL or Backspace.

Also added vertical padding to auto-zoom, to be able to see low/high
labels drawn in the chart.