Correct real-time updates of a data source
The X-range limitations were updated incorrectly across linked plots when
one of the data sources were updated with new data at the far right of the
plot.

Symbol (scatter) plot now uses the last line plot color by default, if
none is specified. So if you, without specifying color, draw a 1st line
and a symbol for that, and then another line and another symbol, the first
two get one color and the next two another according to the principle of
least astonishment.