Allow plotting of already indexed series
If you insert data into the same dataframe as you already supplied for
another plot (say you've already plotted candlestick and now add a MA),
you don't have to pass on time data, the index is enough.

Also made it possible to reach all pixels in the plot with the crosshair
(when moving mouse slowly). Before it was not always possible due to
rounding of the scene Y-axis.