Disable scrolling of the plots during mouseLeftDrag
This commit disables the automatic scrolling of the chart while we drag the chart with mouse left button. This allows users to smoothly drag the chart left and right without any automatic scrolling interference during the drag operation.

I had noticed that in realtime chart, when I was at far right I couldn't drag the chart to the right slowly because as soon as new data arrived the plots scrolled to the left automatically. I could only drag the chart to the right when I did so quickly.